flexloader/kafkatoadb/kafktoadb/variant-9-k8s-operator/configs/requirements.txt
# Основные зависимости
pyyaml>=6.0.1
psutil>=5.9.5
python-dateutil>=2.8.2
typing-extensions>=4.5.0

# Зависимости для работы с Kafka
kafka-python>=2.0.2
confluent-kafka>=2.2.0

# Зависимости для работы с ADB
pymysql>=1.1.0
sqlalchemy>=2.0.15

# Зависимости для мониторинга и метрик
prometheus-client>=0.17.1
statsd>=4.0.1

# Зависимости для безопасности
cryptography>=41.0.1
certifi>=2023.7.22

# Зависимости для тестирования
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Зависимости для разработки
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1 